backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Redis
redis==5.2.1

# Anthropic Claude
anthropic==0.43.0

# MCP
mcp==1.3.0

# OpenTelemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-redis==0.50b0

# Security
PyNaCl==1.5.0

# Utilities
structlog==24.4.0
tenacity==9.0.0
